Corrected nre project pages paths.
