+++ lib/pages/projects/classic_snake_ffps/classic_snake_ffps.dart
@@ -1,16 +1,15 @@
 import 'package:portfolio_website/components/projects/project_component.dart';
 import 'package:portfolio_website/components/projects/text_section.dart';
-import 'package:portfolio_website/pages/project_page/project_page_template.dart';
+import 'package:portfolio_website/pages/projects/templates/project_page_template.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-
 import '../../../providers/project_page_provider.dart';
 
-class MachineLearning extends StatelessWidget {
-  MachineLearning({super.key});
+class ClassicSnakeFFPS extends StatelessWidget {
+  ClassicSnakeFFPS({super.key});
 
   final List<TextSection> textSections = [
-    TextSection(title: "Description", bodyPath: "This is a project for machine learning"),
+    TextSection(title: "Description", bodyPath: "This is a project for snake"),
     TextSection(title: "Description", bodyPath: "This is a project"),
   ];
   final List<ProjectComponent> projectComponents = [ProjectComponent(title: "Description", subComponents: [])];
@@ -18,10 +17,10 @@ class MachineLearning extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     ProjectPageProvider projectPageProvider = Provider.of<ProjectPageProvider>(context, listen: false);
-    Project project = projectPageProvider.getProject("Machine Learning");
+    Project project = projectPageProvider.getProject("Classic Snake Focused Frames per Second");
     return ProjectPageTemplate(
       projectTitle: project.title,
-      bannerImage: "lib/assets/images/project_images/machine_learning.png",
+      bannerImage: "lib/assets/images/project_images/snake.png",
       textSections: textSections,
       projectComponents: projectComponents,
     );
