+++ lib/pages/projects/data_into_ipl/data_into_ipl.dart
@@ -1,16 +1,16 @@
 import 'package:portfolio_website/components/projects/project_component.dart';
 import 'package:portfolio_website/components/projects/text_section.dart';
-import 'package:portfolio_website/pages/project_page/project_page_template.dart';
+import 'package:portfolio_website/pages/projects/templates/project_page_template.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import '../../../providers/project_page_provider.dart';
 
-class LetsTalkk extends StatelessWidget {
-  LetsTalkk({super.key});
+class DataIntoIPL extends StatelessWidget {
+  DataIntoIPL({super.key});
 
   final List<TextSection> textSections = [
-    TextSection(title: "Description", bodyPath: "This is a project for trivago for games"),
+    TextSection(title: "Description", bodyPath: "This is a project for IPL"),
     TextSection(title: "Description", bodyPath: "This is a project"),
   ];
   final List<ProjectComponent> projectComponents = [ProjectComponent(title: "Description", subComponents: [])];
@@ -18,10 +18,10 @@ class LetsTalkk extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     ProjectPageProvider projectPageProvider = Provider.of<ProjectPageProvider>(context, listen: false);
-    Project project = projectPageProvider.getProject("LetsTalkk");
+    Project project = projectPageProvider.getProject("Data Into IPL");
     return ProjectPageTemplate(
       projectTitle: project.title,
-      bannerImage: "lib/assets/images/project_images/messaging.png",
+      bannerImage: "lib/assets/images/project_images/data_in_ipl.png",
       textSections: textSections,
       projectComponents: projectComponents,
     );
