+++ lib/pages/projects/hospital_management_utility/hospital_management_utility.dart
@@ -1,16 +1,15 @@
 import 'package:portfolio_website/components/projects/project_component.dart';
 import 'package:portfolio_website/components/projects/text_section.dart';
-import 'package:portfolio_website/pages/project_page/project_page_template.dart';
+import 'package:portfolio_website/pages/projects/templates/project_page_template.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-
 import '../../../providers/project_page_provider.dart';
 
-class PortfolioWebsite extends StatelessWidget {
-  PortfolioWebsite({super.key});
+class HospitalManagementUtility extends StatelessWidget {
+  HospitalManagementUtility({super.key});
 
   final List<TextSection> textSections = [
-    TextSection(title: "Description", bodyPath: "This is a project for my portfolio"),
+    TextSection(title: "Description", bodyPath: "This is a project for hospitals"),
     TextSection(title: "Description", bodyPath: "This is a project"),
   ];
   final List<ProjectComponent> projectComponents = [ProjectComponent(title: "Description", subComponents: [])];
@@ -18,10 +17,10 @@ class PortfolioWebsite extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     ProjectPageProvider projectPageProvider = Provider.of<ProjectPageProvider>(context, listen: false);
-    Project project = projectPageProvider.getProject("Portfolio Website");
+    Project project = projectPageProvider.getProject("Hospital Management Utility");
     return ProjectPageTemplate(
       projectTitle: project.title,
-      bannerImage: "lib/assets/images/project_images/portfolio.png",
+      bannerImage: "lib/assets/images/project_images/hospital_management_utility.png",
       textSections: textSections,
       projectComponents: projectComponents,
     );
